+++ .travis.yml
@@ -25,8 +25,10 @@ jobs:
         - docker
       before_install:
         - docker build -t buzz/tests .
-        - docker run -d -p 127.0.0.1:8080:80 -p 127.0.0.1:10000:10000 -p 127.0.0.1:3128:3128 buzz/tests
+        - docker run -d -p 127.0.0.1:8080:80 -p 127.0.0.1:3128:3128 buzz/tests
         - docker ps -a
+      before_script:
+        - ./vendor/bin/http_test_server > /dev/null 2>&1 &
       script:
         - ./vendor/bin/phpunit --testsuite Integration
