+++ tests/Integration/Httplug/BaseIntegrationTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Buzz\Test\Integration;
+namespace Buzz\Test\Integration\Httplug;
 
 use Http\Client\Tests\HttpClientTest;
