+++ tests/Integration/Httplug/CurlIntegrationTest.php
@@ -2,18 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Buzz\Test\Integration;
+namespace Buzz\Test\Integration\Httplug;
 
-use Buzz\Client\MultiCurl;
+use Buzz\Client\Curl;
 use Nyholm\Psr7\Factory\Psr17Factory;
 
-class MultiCurlIntegrationTest extends BaseIntegrationTest
+class CurlIntegrationTest extends BaseIntegrationTest
 {
     protected function createHttpAdapter()
     {
-        $client = new MultiCurl(new Psr17Factory(), []);
-
-        return $client;
+        return new Curl(new Psr17Factory(), []);
     }
 
     /**
@@ -40,6 +38,7 @@ class MultiCurlIntegrationTest extends BaseIntegrationTest
         if (null !== $body && '1.0' !== $protocolVersion) {
             $this->markTestSkipped('cURL can not send body using GET');
         }
+
         parent::testSendRequestWithOutcome($uriAndOutcome, $protocolVersion, $headers, $body);
     }
 }
