+++ tests/Integration/Httplug/FileGetContentsIntegrationTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Buzz\Test\Integration;
+namespace Buzz\Test\Integration\Httplug;
 
 use Buzz\Client\FileGetContents;
 use Nyholm\Psr7\Factory\Psr17Factory;
@@ -11,8 +11,6 @@ class FileGetContentsIntegrationTest extends BaseIntegrationTest
 {
     protected function createHttpAdapter()
     {
-        $client = new FileGetContents(new Psr17Factory(), []);
-
-        return $client;
+        return new FileGetContents(new Psr17Factory(), []);
     }
 }
